backend/requirements.txt
Django
djangorestframework
psycopg
pillow